# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.DestinationCreateRequest:
    properties:
      detail_alamat:
        type: string
      judul_alamat:
        type: string
      nama_penerima:
        type: string
      no_telp:
        type: string
    required:
    - detail_alamat
    - judul_alamat
    - nama_penerima
    - no_telp
    type: object
  models.DestinationUpdateRequest:
    properties:
      detail_alamat:
        type: string
      nama_penerima:
        type: string
      no_telp:
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@example.com
    name: API Support
  description: PBI API Documentation
  termsOfService: http://swagger.io/terms/
  title: PBI API
  version: "1.0"
paths:
...
  /user/alamat:
    get:
      description: Get all destination addresses owned by logged-in user
      produces:
      - application/json
      responses:
        "200":
          description: Success get destination list
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get all destinations
      tags:
      - Destination
    post:
      consumes:
      - application/json
      description: Create new destination address for logged-in user
      parameters:
      - description: Destination payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.DestinationCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success create destination
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Create destination address
      tags:
      - Destination
  /user/alamat/{id}:
    delete:
      description: Delete destination address by ID (owned by user)
      parameters:
      - description: Destination ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success delete destination
          schema:
            type: object
        "400":
          description: Invalid destination ID
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Destination not found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Delete destination
      tags:
      - Destination
    get:
      description: Get destination address detail by ID (owned by user)
      parameters:
      - description: Destination ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success get destination detail
          schema:
            type: object
        "400":
          description: Invalid destination ID
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Destination not found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get destination by ID
      tags:
      - Destination
    put:
      consumes:
      - application/json
      description: Update destination address by ID (owned by user)
      parameters:
      - description: Destination ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.DestinationUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success update destination
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "404":
          description: Destination not found
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Update destination
      tags:
      - Destination
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"